moves get_raw_future_blocks to cached_state.rs
